Chapter 3: make an example more clear
It took me a minute to understand it was salad that was going to
be thrown out, and possibly because I already know how C<pop>
works.

Wikipedia kind of agrees it is one, but not really that much more
than the tomatoes in the others.

Replacing it with a more generally-hated vegetable.