Added <> to readline sentence.  Warning: might break POD parsing.
When reading the sentence this commit changes, I briefly wondered what the
readline operator is, and then I realized that it is the <> operator.  The <>
operator is used in the example.  Also, earlier in this chapter you mention the
topic variable, and add a '(C<$_>)' after it, so I just added the same thing
here.

perlpod says there are 2 ways to do this.  You can use E<> with special codes,
which might not work in Pod::PseudoPod, or you can use C<< <> >> as I did in
this commit.  See perlpod for more.  Also, I can't get Pod::PseudoPod to work on
my computer, so I was unable to actually test if this commit works or not, but
it should, so I am commiting it with a warning.  Just check this commit if you
cherry-pick it.