Replaced confusing big word "penultimate."
I also replaced "final" with "last," just because I replace "penultimate" with
"second to last," and so I changed "final" to "last" just to make the paragraph
more cohesive and similar.

I think you should make this change, because this is non-fiction.  Your goal is
to explain something to people who don't know it, and using big words just gets
in the way.